+++ test/config/exotic.js
@@ -1,0 +1,1 @@
+module.exports = { when: new Date(0), m: new Map([[1, 2]]), re: /x/, plain: { n: 1 } }