+++ test/config/proto.js
@@ -1,0 +1,1 @@
+module.exports = JSON.parse('{"a":{"__proto__":{"polluted":1}},"__proto__":{"polluted":1}}')